Add new ICMP packet analyzer, remove old one
