+++ src/analyzer/protocol/CMakeLists.txt
@@ -14,7 +14,6 @@ add_subdirectory(gnutella)
 add_subdirectory(gssapi)
 add_subdirectory(gtpv1)
 add_subdirectory(http)
-add_subdirectory(icmp)
 add_subdirectory(ident)
 add_subdirectory(imap)
 add_subdirectory(irc)
