--- src/analyzer/protocol/icmp/CMakeLists.txt
@@ -1,9 +1,0 @@
-
-include(ZeekPlugin)
-
-include_directories(BEFORE ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR})
-
-zeek_plugin_begin(Zeek ICMP)
-zeek_plugin_cc(ICMP.cc Plugin.cc)
-zeek_plugin_bif(events.bif)
-zeek_plugin_end()
